+++ Assets/Scripts/CompleteLevel.cs
@@ -1,0 +1,24 @@
+using UnityEditor.SearchService;
+using UnityEngine;
+
+public class CompleteLevel : MonoBehaviour
+{
+    public string menuSceneName = "MainMenu";
+
+    public string nextLevel = "Level02";
+    public int levelToUnlock = 2;
+
+    public SceneFader sceneFader;
+
+    public void Continue()
+    {
+        Debug.Log("LEVEL WON!");
+        PlayerPrefs.SetInt("levelReached", levelToUnlock);
+        sceneFader.FadeTo(nextLevel);
+    }
+
+    public void Menu()
+    {
+        sceneFader.FadeTo(menuSceneName);
+    }
+}
