+++ Assets/Scripts/GameOver.cs
@@ -1,19 +1,11 @@
 using UnityEngine;
-using TMPro;
 using UnityEngine.SceneManagement;
 
 public class GameOver : MonoBehaviour
 {
-    public TMP_Text roundsText;
-
     public string menuSceneName = "MainMenu";
 
     public SceneFader sceneFader;
-
-    void OnEnable()
-    {
-        roundsText.text = PlayerStats.Rounds.ToString();
-    }
 
     public void Retry()
     {
